feat: refactor the object model and covert with unit tests
Refs: #312
