+++ src/main/java/ch/sbb/polarion/extension/pdf_exporter/rest/controller/ConverterApiController.java
@@ -11,7 +11,6 @@ import ch.sbb.polarion.extension.pdf_exporter.rest.model.conversion.ExportParams
 import ch.sbb.polarion.extension.pdf_exporter.rest.model.conversion.Orientation;
 import ch.sbb.polarion.extension.pdf_exporter.rest.model.conversion.PaperSize;
 import ch.sbb.polarion.extension.pdf_exporter.service.PdfExporterPolarionService;
-import ch.sbb.polarion.extension.pdf_exporter.util.DocumentDataHelper;
 import ch.sbb.polarion.extension.pdf_exporter.util.PdfValidationService;
 import org.jetbrains.annotations.VisibleForTesting;
 import org.springframework.web.context.request.RequestAttributes;
@@ -33,14 +32,15 @@ public class ConverterApiController extends ConverterInternalController {
 
     @VisibleForTesting
     @SuppressWarnings("squid:S5803")
-    ConverterApiController(PdfExporterPolarionService pdfExporterPolarionService,
-                           PdfConverter pdfConverter,
-                           PdfValidationService pdfValidationService,
-                           DocumentDataHelper documentDataHelper,
-                           PdfConverterJobsService pdfConverterJobService,
-                           UriInfo uriInfo,
-                           HtmlToPdfConverter htmlToPdfConverter) {
-        super(pdfExporterPolarionService, pdfConverter, pdfValidationService, documentDataHelper, pdfConverterJobService, uriInfo, htmlToPdfConverter);
+    ConverterApiController(
+            PdfExporterPolarionService pdfExporterPolarionService,
+            PdfConverter pdfConverter,
+            PdfValidationService pdfValidationService,
+            PdfConverterJobsService pdfConverterJobService,
+            UriInfo uriInfo,
+            HtmlToPdfConverter htmlToPdfConverter
+    ) {
+        super(pdfConverter, pdfValidationService, pdfConverterJobService, uriInfo, htmlToPdfConverter);
         this.polarionService = pdfExporterPolarionService;
     }
